+++ R/poisson.R
@@ -1,6 +1,6 @@
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Generate poisson disc samples
+#' Generate Poisson disk samples in 2D
 #'
 #' @param w,h width and height of region
 #' @param r minimum distance between points
@@ -10,19 +10,19 @@
 #' @return data.frame with x and y coordinates and the 'idx' order in which
 #'         the points were added.
 #' @examples
-#' pts <- poisson2d(w = 10, h = 10, r = 2)
-#' plot(pts)
+#' pts <- poisson2d(w = 40, h = 40, r = 1)
+#' plot(pts, asp = 1, ann = FALSE, axes = FALSE, pch = 19)
 #' @importFrom stats runif
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-poisson2d <- function(w = 400, h = 300, r = 10, k = 30L, verbosity = 0L) {
+poisson2d <- function(w = 10, h = 10, r = 2, k = 30L, verbosity = 0L) {
  .Call(poisson2d_, w, h, r, k, verbosity) 
 }
 
 
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#' Generate poisson disc samples in 3D
+#' Generate Poisson disk samples in 3D
 #'
 #' @param w,h,d width and height and depth of region
 #' @param r minimum distance between points
@@ -37,7 +37,7 @@ poisson2d <- function(w = 400, h = 300, r = 10, k = 30L, verbosity = 0L) {
 #' @importFrom stats runif
 #' @export
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-poisson3d <- function(w = 400, h = 300, d = 100, r = 10, k = 30L, verbosity = 0L) {
+poisson3d <- function(w = 10, h = 10, d = 10, r = 4, k = 30L, verbosity = 0L) {
   .Call(poisson3d_, w, h, d, r, k, verbosity) 
 }